+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- [**BC**] Make `Authorize.authorizeAction` function really generic by adding a `formatError` function
+- [**BC**] Rename `Authorize.authorizeAction` to `Authorize.action`
 
 ## 1.0.0 - 2021-01-21
 - Initial implementation